css/css-values/getComputedStyle-calc-background-position-percent-012.html

<!DOCTYPE html>

  <meta charset="UTF-8">

  <title>CSS Values Test: computed value of background-position: calc(100% - 100% + 1em)</title>

  <!--
  "
  Where percentages are not resolved at computed-value time,
  they are not resolved in math functions, e.g.
  'calc(100% - 100% + 1px)' resolves to 'calc(0% + 1px)', not to
  '1px'. If there are special rules for computing percentages
  in a value (e.g. the 'height' property), they apply whenever
  a math function contains percentages.
  "
  § 10.11 Computed Value
  https://www.w3.org/TR/css-values-4/#calc-computed-value
  -->

  <link rel="author" title="Gérard Talbot" href="http://www.gtalbot.org/BrowserBugsSection/css21testsuite/">
  <link rel="help" href="https://www.w3.org/TR/css-values-4/#calc-computed-value">

  <meta name="flags" content="">
  <meta content="This test verifies that a term with a percentage unit that can not be resolved at computed-value time will require a calc() wrapper. And, in such case, even if it has the 0 value, it must be preserved. A term with an em value, on the other hand, can be resolved at computed-value time and therefore must be absolutized to 'px'." name="assert">

  <script src="/resources/testharness.js"></script>

  <script src="/resources/testharnessreport.js"></script>

  <style>
  div#target
    {
      background-color: yellow;
      background-image: url("support/cat.png");
      background-repeat: no-repeat;
      background-size: 14% 50%; /* entirely arbitrary and random background-size values */
      font-size: 16px;
      height: 200px;
    }
  </style>

  <div id="target"></div>

  <script>
  function startTesting()
  {

  var targetElement = document.getElementById("target");

    function verifyComputedStyle(property_name, specified_value, expected_value, description)
    {

    test(function()
      {

      targetElement.style.setProperty(property_name, specified_value);

      assert_equals(getComputedStyle(targetElement)[property_name], expected_value);

      }, description);
    }

    verifyComputedStyle("background-position", "calc(100% - 100% + 1em)", "calc(0% + 16px) 50%", "testing background-position: calc(100% - 100% + 1em)");

    /*
    "Where percentages are not resolved at computed-value time,
    they are not resolved in math functions (...)"
    https://www.w3.org/TR/css-values-4/#calc-computed-value

    Therefore here, the percentage is preserved and
    a calc() wrapper must be used. The 1em term
    must be resolved though.
    */

  }

  startTesting();

  </script>
